requirements.txt
datasets==2.18.0
einops==0.7.0
fsspec==2024.2.0
git-lfs==1.6
h5py==3.10.0
hydra-core==1.3.2
ipdb==0.13.13
lightning==2.2.1
notebook==7.1.1
nvitop==1.3.2
omegaconf==2.3.0
packaging==23.2
pandas==2.2.1
rich==13.7.1
seaborn==0.13.2
scikit-learn==1.4.0
timm==0.9.16
transformers==4.38.2
triton==2.2.0
wandb==0.13.5
